+++ src/main/java/me/gg/pinit/pinittask/application/statistics/event/handler/ScheduleCanceledEventHandler.java
@@ -2,9 +2,11 @@ package me.gg.pinit.pinittask.application.statistics.event.handler;
 
 import me.gg.pinit.pinittask.application.schedule.service.ScheduleService;
 import me.gg.pinit.pinittask.application.statistics.service.StatisticsService;
+import me.gg.pinit.pinittask.application.task.service.TaskService;
 import me.gg.pinit.pinittask.domain.schedule.event.ScheduleCanceledEvent;
 import me.gg.pinit.pinittask.domain.schedule.model.CompletedState;
 import me.gg.pinit.pinittask.domain.schedule.model.Schedule;
+import me.gg.pinit.pinittask.domain.task.model.Task;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
@@ -14,19 +16,25 @@ import org.springframework.transaction.event.TransactionalEventListener;
 public class ScheduleCanceledEventHandler {
     private final StatisticsService statisticsService;
     private final ScheduleService scheduleService;
+    private final TaskService taskService;
 
-    public ScheduleCanceledEventHandler(StatisticsService statisticsService, ScheduleService scheduleService) {
+    public ScheduleCanceledEventHandler(StatisticsService statisticsService, ScheduleService scheduleService, TaskService taskService) {
         this.statisticsService = statisticsService;
         this.scheduleService = scheduleService;
+        this.taskService = taskService;
     }
     @Async
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     public void on(ScheduleCanceledEvent event) {
         if (!event.getBeforeState().equals(CompletedState.COMPLETED)) return;
         Schedule schedule = scheduleService.getSchedule(event.getOwnerId(), event.getScheduleId());
+        if (schedule.getTaskId() == null) {
+            return;
+        }
+        Task task = taskService.getTask(event.getOwnerId(), schedule.getTaskId());
         statisticsService.removeElapsedTime(
                 event.getOwnerId(),
-                schedule.getTemporalConstraint().getTaskType(),
+                task.getTemporalConstraint().getTaskType(),
                 schedule.getHistory().getElapsedTime(),
                 schedule.getDesignatedStartTime()
         );
